refactored frontend (#15)
