+++ frontend/src/app/Auth/SignUp/components(redux)/SignupForm.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { useSelector } from 'react-redux';
-import { RootState } from '@/redux/store';
+import { RootState } from '@/store/store';
 import StepOne from './StepOne';
 import StepTwo from './StepTwo';
 import StepThree from './StepThree';
