+++ frontend/src/app/Auth/SignUp/layout.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { Provider } from "react-redux";
-import { store } from '@/redux/store';
+import { store } from '@/store/store';
 
 export default function SignupLayout({
