--- frontend/src/app/Dashboard/Network/page.tsx
@@ -1,14 +1,0 @@
-import DashboardNavBar from "../Components/DashboardNavbar"
-
-export default function Network() {
-    return (
-        <div>
-            <DashboardNavBar />
-            <div className="flex flex-col items-center justify-center min-h-screen"> 
-                <h1 className="text-3xl">Network PAGE </h1> 
-                <p>Still under construction! Check back later</p> 
-                <p>click the recover logo to return to the homepage</p>
-            </div>
-        </div>
-    )
-}
