+++ frontend/src/app/Dashboard/projects/page.tsx
@@ -1,11 +1,11 @@
-import DashboardNavBar from "../../features/dashboard/DashboardNavbar"
+import DashboardNavBar from "../components/DashboardNavbar"
 
-export default function Project() {
+export default function AllProjects() {
     return (
         <div>
-            <DashboardNavBar />
+            <DashboardNavBar/>
             <div className="flex flex-col items-center justify-center min-h-screen"> 
-                <h1 className="text-3xl">Dashboard </h1> 
+                <h1 className="text-3xl">ALL PROJECTS </h1> 
                 <p>Still under construction! Check back later</p> 
                 <p>click the recover logo to return to the homepage</p>
             </div>
