--- frontend/src/app/User/page.tsx
@@ -1,10 +1,0 @@
-export default function User() {
-    return (
-        <div className="flex flex-col items-center justify-center min-h-screen"> 
-            <h1 className="text-3xl">USER DETAILS </h1> 
-            <p>Still under construction! Check back later</p> 
-            <p>click the recover logo to return to the homepage</p>
-            <p>Not logged in so no information</p>
-        </div>
-    )
-}
