+++ frontend/src/components/ui/eyePasswordIcon.tsx
