+++ frontend/src/hooks/common/useDebounce.ts
@@ -1,0 +1,15 @@
+import { useEffect, useState } from "react";
+
+export const useDebounce = <T>(value: T, delay: number = 500): T => {
+    const [debouncedValue, setDebouncedValue] = useState<T>(value);
+  
+    useEffect(() => {
+      const timer = setTimeout(() => {
+        setDebouncedValue(value);
+      }, delay);
+  
+      return () => clearTimeout(timer);
+    }, [value, delay]);
+  
+    return debouncedValue;
+  };
