+++ frontend/src/hooks/common/useOutsideClick.ts
@@ -1,0 +1,18 @@
+import { useEffect, useRef } from "react";
+
+export const useOutsideClick = (callback: () => void) => {
+    const ref = useRef<HTMLDivElement>(null);
+  
+    useEffect(() => {
+      const handleClick = (event: MouseEvent) => {
+        if (ref.current && !ref.current.contains(event.target as Node)) {
+          callback();
+        }
+      };
+  
+      document.addEventListener('mousedown', handleClick);
+      return () => document.removeEventListener('mousedown', handleClick);
+    }, [callback]);
+  
+    return ref;
+  };
