+++ frontend/src/store/store.ts
@@ -1,0 +1,127 @@
+// import signupReducer from './slices/auth/signupSlice'
+
+// export const store = configureStore({
+//     reducer: {
+//         signup: signupReducer,
+//     },
+//     middleware: (getDefaultMiddleware) =>
+//         getDefaultMiddleware({
+//             serializableCheck: false,
+//         }),
+// });
+
+// export type RootState = ReturnType<typeof store.getState>;
+// export type AppDispatch = typeof store.dispatch;
+
+// store/store.ts
+import { configureStore } from '@reduxjs/toolkit';
+import { rootReducer } from './rootReducer';
+import { authApi } from './features/auth/authApi';
+import { projectApi } from './features/projects/projectApi';
+
+export const store = configureStore({
+  reducer: rootReducer,
+  middleware: (getDefaultMiddleware) =>
+    getDefaultMiddleware().concat(
+      authApi.middleware,
+      projectApi.middleware
+    ),
+  devTools: process.env.NODE_ENV !== 'production',
+});
+
+export type RootState = ReturnType<typeof store.getState>;
+export type AppDispatch = typeof store.dispatch;
+
+// store/rootReducer.ts
+import { combineReducers } from '@reduxjs/toolkit';
+import { authApi } from './features/auth/authApi';
+import { projectApi } from './features/projects/projectApi';
+import authReducer from './features/auth/authSlice';
+import projectReducer from './features/projects/projectSlice';
+import uiReducer from './features/ui/uiSlice';
+
+export const rootReducer = combineReducers({
+  auth: authReducer,
+  projects: projectReducer,
+  ui: uiReducer,
+  [authApi.reducerPath]: authApi.reducer,
+  [projectApi.reducerPath]: projectApi.reducer,
+});
+
+// store/features/auth/authSlice.ts
+import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
+import type { RootState } from '../../store';
+
+interface AuthState {
+  user: User | null;
+  isLoading: boolean;
+  error: string | null;
+}
+
+const initialState: AuthState = {
+  user: null,
+  isLoading: false,
+  error: null,
+};
+
+export const logout = createAsyncThunk(
+  'auth/logout',
+  async (_, { dispatch }) => {
+    localStorage.removeItem('token');
+    // Additional cleanup
+  }
+);
+
+const authSlice = createSlice({
+  name: 'auth',
+  initialState,
+  reducers: {
+    setUser: (state, action) => {
+      state.user = action.payload;
+    },
+  },
+  extraReducers: (builder) => {
+    builder.addCase(logout.fulfilled, (state) => {
+      state.user = null;
+    });
+  },
+});
+
+export const { setUser } = authSlice.actions;
+export default authSlice.reducer;
+
+// store/features/auth/authApi.ts
+import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
+import type { RootState } from '../../store';
+
+export const authApi = createApi({
+  reducerPath: 'authApi',
+  baseQuery: fetchBaseQuery({
+    baseUrl: '/api/auth',
+    prepareHeaders: (headers, { getState }) => {
+      const token = (getState() as RootState).auth.user?.token;
+      if (token) {
+        headers.set('authorization', `Bearer ${token}`);
+      }
+      return headers;
+    },
+  }),
+  endpoints: (builder) => ({
+    login: builder.mutation<User, LoginCredentials>({
+      query: (credentials) => ({
+        url: 'login',
+        method: 'POST',
+        body: credentials,
+      }),
+    }),
+    signup: builder.mutation<User, SignupCredentials>({
+      query: (credentials) => ({
+        url: 'signup',
+        method: 'POST',
+        body: credentials,
+      }),
+    }),
+  }),
+});
+
+export const { useLoginMutation, useSignupMutation } = authApi;
