+++ frontend/src/styles/constants/RecoverLogo.tsx
