+++ frontend/src/types/project.ts
@@ -1,0 +1,7 @@
+export interface Project {
+    id: string;
+    name: string;
+    description: string;
+    status: 'active' | 'archived';
+    createdAt: string;
+  }
